requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2; python_version >= '3.7'
django==5.0.1; python_version >= '3.10'
django-cors-headers==4.3.1; python_version >= '3.8'
djangorestframework==3.14.0; python_version >= '3.6'
numpy==1.26.3; python_version >= '3.9'
opencv-python==4.9.0.80; python_version >= '3.6'
packaging==23.2; python_version >= '3.7'
pillow==10.2.0; python_version >= '3.8'
pyjwt==2.8.0; python_version >= '3.7'
pytesseract==0.3.10; python_version >= '3.7'
pytz==2024.1
sqlparse==0.4.4; python_version >= '3.5'
typing-extensions==4.9.0; python_version < '3.11'
tzdata==2023.4; sys_platform == 'win32'
